requirements.txt
asgiref==3.4.1
boto3==1.26.160
botocore==1.29.160
certifi==2021.5.30
charset-normalizer==2.0.4
Django==3.2.5
django-admin-honeypot==1.1.0
django-admin-honeypot-updated-2021==1.2.0
django-admin-thumbnails==0.2.6
django-ipware==5.0.0
django-session-timeout==0.1.0
django-storages==1.11.1
idna==3.2
jmespath==0.10.0
Pillow==9.5.0
psycopg2==2.9.6
psycopg2-binary==2.9.6
python-dateutil==2.8.2
python-decouple==3.4
pytz==2021.1
requests==2.31.0
s3transfer==0.6.1
six==1.12.0
sqlparse==0.4.1
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.5
